Fix for issue 20
This change actually fixes the old issue 20 for throttle warning:
If P1 - P3 is selected as throttle source a warning is still generated
for the throttle stick instead of P1 - P3 according selection.
This is fixed now and saves additonally about 54 bytes on stock.
Please notice, if you select a output channel as throttle source CH01 -
CH16 it again uses throttle stick as source for throttle warning.