New Turnigy9x transmitter fix; Throttle only
Added support for
#define INPUT_WARNINGS_GENERATE_SIM_DATA

The latests deliveries from 9x transmitters (since 2 years now, 2014) do
not stop sending if no pulses are generated.
This fact totally breaks the nice throttle and switch warning concept,
because even we prevent sending pulses the transmitter module ignores
this, and continous to send the last value or 0 at beginning. And 0
means in the middle of throttle -> 50%.
To cope with this situation, we need a new concept how to handle
throttle and switch warnings.
My solution is, not to stop pulses at all. In warning situation the
input of the appropriate channels are ignored and replaced with a value
which wouldn't generate the warning, e.g. for throttle -100%
Because this new concept is also compatible with the better (older)
transmitter we can enable it by default and let the user choose if he
want's the old handling or new one.

Code cost:
=======

Code cost for ATmega64 is only 14 bytes and 1 byte data.
However the switch warning is not yet implemented, The code cost for
this will be a lot more. Still searching for a better solution.